refactor(golyok): rename GOLYOK to ball
As well as GOLYKISZ and GOLYUTK
