+++ ball_handler.h
@@ -1,5 +1,5 @@
-#ifndef GOLYKISZ_H
-#define GOLYKISZ_H
+#ifndef BALL_HANDLER_H
+#define BALL_HANDLER_H
 
 double next_collision_time(int* collider1, int* collider2);
 void simulate_collision(int collider1, int collider2);
