+++ meson.build
@@ -28,9 +28,9 @@ elma_sources = files(
     'EDITUJ.CPP',
     'ED_CHECK.CPP',
     'FOGOCSKA.CPP',
-    'GOLYKISZ.CPP',
-    'GOLYOK.CPP',
-    'GOLYUTK.CPP',
+    'ball_handler.cpp',
+    'ball.cpp',
+    'ball_collision.cpp',
     'GYURU.CPP',
     'H_WAV.CPP',
     'JATEKOS.CPP',
